+++ noodles-sam/examples/sam_query.rs
@@ -7,10 +7,7 @@
 
 use std::{env, io, path::PathBuf};
 
-use noodles_sam::{
-    self as sam,
-    alignment::{io::Write, RecordBuf},
-};
+use noodles_sam::{self as sam, alignment::io::Write};
 
 const UNMAPPED: &str = "*";
 
@@ -23,8 +20,8 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut reader = sam::indexed_reader::Builder::default().build_from_path(src)?;
     let header = reader.read_header()?;
 
-    let records: Box<dyn Iterator<Item = io::Result<RecordBuf>>> = if raw_region == UNMAPPED {
-        reader.query_unmapped(&header).map(Box::new)?
+    let records: Box<dyn Iterator<Item = io::Result<sam::Record>>> = if raw_region == UNMAPPED {
+        reader.query_unmapped().map(Box::new)?
     } else {
         let region = raw_region.parse()?;
         reader.query(&header, &region).map(Box::new)?