+++ noodles-sam/src/indexed_reader.rs
@@ -89,15 +89,12 @@ where
         &'a mut self,
         header: &'a Header,
         region: &Region,
-    ) -> io::Result<impl Iterator<Item = io::Result<RecordBuf>> + 'a> {
+    ) -> io::Result<impl Iterator<Item = io::Result<Record>> + 'a> {
         self.inner.query(header, &self.index, region)
     }
 
     /// Returns an iterator of unmapped records after querying for the unmapped region.
-    pub fn query_unmapped<'a>(
-        &'a mut self,
-        header: &'a Header,
-    ) -> io::Result<impl Iterator<Item = io::Result<RecordBuf>> + 'a> {
-        self.inner.query_unmapped(header, &self.index)
+    pub fn query_unmapped(&mut self) -> io::Result<impl Iterator<Item = io::Result<Record>> + '_> {
+        self.inner.query_unmapped(&self.index)
     }
 }