+++ noodles-sam/src/reader/query.rs
@@ -1,44 +1,36 @@
 use std::io::{self, Read, Seek};
 
 use noodles_bgzf as bgzf;
-use noodles_core::region::Interval;
+use noodles_core::{region::Interval, Position};
 use noodles_csi::{self as csi, binning_index::index::reference_sequence::bin::Chunk};
 
 use super::Reader;
-use crate::{alignment::RecordBuf, Header};
+use crate::{alignment::Record as _, Header, Record};
 
 pub struct Query<'a, R>
 where
     R: Read + Seek,
 {
     reader: Reader<csi::io::Query<'a, R>>,
-    header: &'a Header,
-    record: RecordBuf,
+    record: Record,
 }
 
 impl<'a, R> Query<'a, R>
 where
     R: Read + Seek,
 {
-    pub(super) fn new(
-        reader: &'a mut bgzf::Reader<R>,
-        header: &'a Header,
-        chunks: Vec<Chunk>,
-    ) -> Self {
+    pub(super) fn new(reader: &'a mut bgzf::Reader<R>, chunks: Vec<Chunk>) -> Self {
         Self {
             reader: Reader::new(csi::io::Query::new(reader, chunks)),
-            header,
-            record: RecordBuf::default(),
+            record: Record::default(),
         }
     }
 
-    fn next_record(&mut self) -> io::Result<Option<RecordBuf>> {
-        self.reader
-            .read_record_buf(self.header, &mut self.record)
-            .map(|n| match n {
-                0 => None,
-                _ => Some(self.record.clone()),
-            })
+    fn next_record(&mut self) -> io::Result<Option<Record>> {
+        self.reader.read_record(&mut self.record).map(|n| match n {
+            0 => None,
+            _ => Some(self.record.clone()),
+        })
     }
 }
 
@@ -46,7 +38,7 @@ impl<'a, R> Iterator for Query<'a, R>
 where
     R: Read + Seek,
 {
-    type Item = io::Result<RecordBuf>;
+    type Item = io::Result<Record>;
 
     fn next(&mut self) -> Option<Self::Item> {
         match self.next_record() {
@@ -57,30 +49,37 @@ where
     }
 }
 
-pub struct FilterByRegion<I> {
+pub struct FilterByRegion<'h, I> {
     records: I,
+    header: &'h Header,
     reference_sequence_id: usize,
     interval: Interval,
 }
 
-impl<I> FilterByRegion<I>
+impl<'h, I> FilterByRegion<'h, I>
 where
-    I: Iterator<Item = io::Result<RecordBuf>>,
+    I: Iterator<Item = io::Result<Record>>,
 {
-    pub fn new(records: I, reference_sequence_id: usize, interval: Interval) -> Self {
+    pub fn new(
+        records: I,
+        header: &'h Header,
+        reference_sequence_id: usize,
+        interval: Interval,
+    ) -> Self {
         Self {
             records,
+            header,
             reference_sequence_id,
             interval,
         }
     }
 }
 
-impl<I> Iterator for FilterByRegion<I>
+impl<'h, I> Iterator for FilterByRegion<'h, I>
 where
-    I: Iterator<Item = io::Result<RecordBuf>>,
+    I: Iterator<Item = io::Result<Record>>,
 {
-    type Item = io::Result<RecordBuf>;
+    type Item = io::Result<Record>;
 
     fn next(&mut self) -> Option<Self::Item> {
         loop {
@@ -89,18 +88,41 @@ where
                 Err(e) => return Some(Err(e)),
             };
 
-            if let (Some(id), Some(start), Some(end)) = (
-                record.reference_sequence_id(),
-                record.alignment_start(),
-                record.alignment_end(),
+            match intersects(
+                self.header,
+                &record,
+                self.reference_sequence_id,
+                self.interval,
             ) {
-                let alignment_interval = Interval::from(start..=end);
-
-                if id == self.reference_sequence_id && self.interval.intersects(alignment_interval)
-                {
-                    return Some(Ok(record));
-                }
+                Ok(true) => return Some(Ok(record)),
+                Ok(false) => {}
+                Err(e) => return Some(Err(e)),
             }
         }
+    }
+}
+
+fn intersects(
+    header: &Header,
+    record: &Record,
+    reference_sequence_id: usize,
+    region_interval: Interval,
+) -> io::Result<bool> {
+    match (
+        record
+            .reference_sequence_id(header)
+            .map(usize::try_from)
+            .transpose()?,
+        record
+            .alignment_start()
+            .map(Position::try_from)
+            .transpose()?,
+        record.alignment_end().transpose()?,
+    ) {
+        (Some(id), Some(start), Some(end)) => {
+            let alignment_interval = (start..=end).into();
+            Ok(id == reference_sequence_id && region_interval.intersects(alignment_interval))
+        }
+        _ => Ok(false),
     }
 }